+++ .gitignore
@@ -2,3 +2,6 @@
 *.idea
 *.vscode
 *.log
+
+# python
+__pycache__
