+++ culte-client/src/assets/castle.png
