--- culte-client/src/utils/utils.js
@@ -1,7 +1,0 @@
-
-function log(message) {
-  const log = document.getElementsByName("log")[0];
-  log.value += message + "\n";
-}
-
-export default log;
